Testing ms timestamp

+++ DAC/src/command/GameplayCommands.java
@@ -3,8 +3,8 @@ package command;
 public interface GameplayCommands extends Command {
     public int getX();
     public int getY();
-    public int getTimeStamp();
-    public void setTimeStamp(int timeStamp);
+    public long getTimeStamp();
+    public void setTimeStamp(long timeStamp);
     public int getConnectionID();
     public void setConnectionID(int connectionID);
 }
